+++ src/types/common.ts
@@ -42,6 +42,9 @@ export const embedsColors: Record<string, string> = {
   Command: "#62A1FE",
   Game: "#FFAD83",
   Wallet: "#5CD97D",
+  Payme: "#FFD652",
+  Paylink: "#FF9900",
+  PayStatusList: "#36CFFF",
 }
 
 export type SlashCommandChoiceOption = {